requirements.txt
click==6.6
Django==1.10.2
django-filter==0.15.2
djangorestframework==3.4.7
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.7
MarkupSafe==0.23
PyWit==0.4.0
requests==2.11.1
Werkzeug==0.11.11
